Add Book ("Estratégia" de Investimento)
A book is like a category of investment. It helps you visualize the
allocation strategy of your current portfolio. Examples of books: Stocks,
Options, Gold, Long/Short, Buy and Hold, Foreign Exchange. A book may belong
to another book, so you can build a collection of books like the following:

> Renda Fixa
    > CDI
    > IPCA
    > Préfixado
> Ações
> Ouro

In this tree example, "CDI", "IPCA" and "Préfixado" belong to "Renda Fixa".

The application will not work with a tree with level >= 3 and it will not
validate that a Book is being created in such levels. The GUI will not allow
such behaviour.

When a new user is created, a collection of default books will be added to
the user's account. It can be changed later by the user with whatever she
thinks is best for organizing her portfolio.

Commands:

- rails generate model Book name:string user:belongs_to position:integer
- rails generate scaffold_controller Book